pyproject.toml
[project]
name = "agentx"
version = "0.1.0"
description = "Multi-agent ML competition solver using A2A protocol"
requires-python = ">=3.11"
dependencies = [
    "a2a-sdk[http-server]>=0.3.0",
    "uvicorn>=0.27.0",
    "openai>=1.50.0",
    "httpx>=0.27.0",
    "numpy>=1.26.0",
    "pandas>=2.1.0",
    "scikit-learn>=1.3.0",
    "lightgbm>=4.1.0",
    "xgboost>=2.0.0",
    "catboost>=1.2.0",
    "optuna>=3.4.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]
